+++ src/inspector/components/LeftColumn/LeftColumnContainer.tsx
@@ -2,7 +2,7 @@ import { connect, MapDispatchToPropsFunction } from "react-redux";
 import { ILeftColumnDispatch, ILeftColumnProps, LeftColumn } from "./LeftColumn";
 import { IRootState } from "../../../shared/store";
 import { getTargetReference, getAutoUpdate, getAddAllowed, getPropertySettings, getLockedSelection, getPinnedSelection, getRemovableSelection, getSelectedDescriptorsUUID, getActiveTargetReference, getActiveTargetDocument, getActiveTargetLayer, getActiveReferenceChannel, getActiveReferenceGuide, getActiveReferencePath, getActiveReferenceActionSet, getActiveReferenceActionItem, getActiveReferenceCommand, getActiveReferenceProperty, getSelectedTargetReference, getDescriptorsListView, getHasAutoActiveDescriptor, getFilterBySelectedReferenceType, getActiveTargetReferenceForAM, getActiveReferenceHistory, getActiveReferenceSnapshot, getInspectorSettings, getActiveTargetReferenceListenerCategory, getSelectedDescriptors, getReplayEnabled } from "../../selectors/inspectorSelectors";
-import { setTargetReferenceAction, addDescriptorAction, setSelectedReferenceTypeAction, clearAction, pinDescAction, removeDescAction, lockDescAction, setFilterStateAction, setListenerAction, setAutoInspectorAction, setSearchTermAction, setRenameModeAction, selectDescriptorAction } from "../../actions/inspectorActions";
+import { setTargetReferenceAction, addDescriptorAction, setSelectedReferenceTypeAction, clearAction, pinDescAction, removeDescAction, lockDescAction, setFilterStateAction, setListenerAction, setAutoInspectorAction, setSearchTermAction, setRenameModeAction, selectDescriptorAction, setDontShowMarketplaceInfoAction } from "../../actions/inspectorActions";
 import { TDocumentReference, TLayerReference, TChannelReference, TGuideReference, TPathReference, TActionSet, TActionItem, TActionCommand, IContentWrapper, TBaseProperty, IHistory, THistoryReference, TSnapshotReference, TListenerCategoryReference, TSelectDescriptorOperation } from "../../model/types";
 
 
@@ -57,7 +57,9 @@ const mapDispatchToProps: MapDispatchToPropsFunction<ILeftColumnDispatch, Record
 		setAutoInspector: (enabled) => dispatch(setAutoInspectorAction(enabled)),
 		setSearchTerm: (str) => dispatch(setSearchTermAction(str)),
 		setRenameMode: (uuid: string, on: boolean) => dispatch(setRenameModeAction(uuid, on)),
-		onSelect: (operation: TSelectDescriptorOperation,uuid?:string) => dispatch(selectDescriptorAction(operation, uuid)),
+		onSelect: (operation: TSelectDescriptorOperation, uuid?: string) => dispatch(selectDescriptorAction(operation, uuid)),
+		
+		onSetDontShowMarketplaceInfo:(enabled:boolean)=>dispatch(setDontShowMarketplaceInfoAction(enabled)),
 	};
 };